Add Turbo Stream actions to add and remove CSS classes
A common UI pattern is to indicate whether a field is valid or
invalid by adding a CSS class to invalid fields to turn them red.
Another is to submit the input as the user enters the text rather
than waiting for them to press a button to submit the form.

Combining the two requires us to add or remove fields from a field
as the user continues to type into it.

This is difficult to implement using Turbo Streams, because the
only action that can change the CSS class of an element on a page
is `update`, which removes the existing element and replaces it
with a new one.  This interrupts the user's input and is likely
to discard their latest changes.

It is possible to work around this, for example by updating a
hidden element with a Stimulus controller that adds or removes the
class; or perhaps doing something with sibling selectors; but this
is bespoke and unwieldy.

The Turbo handbook says that...

> Turbo Streams consciously restricts you to seven actions
> [...in order to...]
> keep you from turning individual responses in a jumble of
> behaviors that cannot be reused and which make the app hard to
> follow

...but I feel that the omission of an action which enables this
behaviour has pushed us in that direction instead of away from it.

So, to enable these sorts of update using Turbo Streams out of the
box, this commit adds two new stream actions: `add-class`, which
adds the template's classes to the targets; and `remove-class`,
which removes them.

Since it's also fairly common to enable and disable elements, I did
consider adding more generic actions to handle element attributes,
but since disabling a field would also interrupt user input I think
it is sufficiently covered by existing actions, and so it's not
worth the extra complexity.