Always accept Turbo Stream responses
Previously, Turbo Streams could only be used with non-GET form
submissions. However there are also cases where it's useful to have a
Turbo Stream response to a GET action.

With this change, all Turbo-initiated requests will include the Turbo
Stream type in the `Accept` header, so that Turbo Streams can be used
with any of them.