Lazy load Frame styled with `display: contents`
Reproduces [hotwired/turbo#625][]

When a `<turbo-frame>` element is styled as [`display:
contents`](https://developer.mozilla.org/en-US/docs/Web/CSS/display#box),
it derives its box model from its descendants, and browsers treat it as
if the element itself is not present in the document.

Unfortunately, [since it's "omitted" from the
document](https://stackoverflow.com/q/72741599), the underlying
[IntersectionObserver](https://developer.mozilla.org/en-US/docs/Web/API/Intersection_Observer_API)
instance that powers the
[`[loading="lazy"]`](https://turbo.hotwired.dev/reference/frames#html-attributes)
behavior never detects when the `<turbo-frame>` contents enter the
viewport.

[hotwired/turbo#625]: https://github.com/hotwired/turbo/issues/625