Introduce client-side Turbo Cache Control API
This commit introduces a `Cache` class which exposes an API that can be
used to control the value of the `turbo-cache-control` meta element
form the client-side.

Most of the times you just need to the set the value of the
`turbo-cache-control` meta element from the server-side when you are
responding to your regular HTML request.

But there are some cases where you want to control the cache control
behavior from the client-side after a certain action happened on the
client (i.e. if you don't want to show remaining artefacts of an
finsihed animation).

You can achieve the same behavior by dynamically prepending a `<meta>`
element to the `<head>` via JavaScript, like:

```javascript
document.head.insertAdjacentHTML(
  'beforeend',
  '<meta name="turbo-cache-control" content="no-cache">'
)
```

Though, this feels very error-prone and doesn't account for duplicate
`<meta>` elements in the `<head>`.

This Pull Request aims to simplify this by exposing a client-side API to
control the value of the meta element.

The name of the exposed functions are kept in-line with the helpers in:
https://github.com/hotwired/turbo-rails/blob/43bf84a9b1d2cc78da6810b82f92be2c32c9dbfd/app/helpers/turbo/drive_helper.rb#L15