Convert search parameters to form inputs when using `data-turbo-stream`
Links with `data-turbo-stream` are converted to forms, and forms with
method="GET" don't submit the query string part of their action. This
can be circumvented by iterating through any query string parameters and
adding them to the generated form as hidden inputs.

Note: MDN refers to what we usually call "query string parameters" as
"search parameters". This is a bit confusing, but it's the correct
terminology I guess :)