Prevent form links from submitting multiple requests
The `LinkInterceptor` class contains some Frame-specific checks that
were at the root of `<a data-turbo-method="...">` element clicks
submitting multiple HTTP requests.

To resolve the underlying issue, this commit changes the
`FormLinkInterceptor` class to rely on an instance of
`LinkClickObserver` instead of `LinkInterceptor`.

In order to make that possible, this commit also extends the
`LinkClickObserver` to accept a second argument to server as the
[EventTarget][] for the `click` event listeners. As a consumer, the
`FormLinkInterceptor` instances attached to `<turbo-frame>` element
scope their listeners to the element, while the `Session`'s
`FormLinkInterceptor` instance attaches a catch-all listener to the
`window`.

Since this commit changes the underlying event listening mechanism, it
also renames the `FormLinkInterceptor` and delegate to
`FormLinkClickObserver` to match the `LinkClickObserver` naming
patterns.

[EventTarget]: https://developer.mozilla.org/en-US/docs/Web/API/EventTarget