Encode Form Submitter `[name]` into submission
Closes [hotwired/turbo#272][]

Do not exclude a `<form>` submitter's `[name]` attribute from the
encoded `FormData` instance, even if a `[value]` is not specified.
Instead, encode the value as an empty string if a `[value]` attribute is
unspecified.

The existing tests included a [button][] to exercise this exclusion
behavior. Their assertions were incorrect, so this commit changes them
to reflect the proper behavior.

[button]: https://github.com/hotwired/turbo/blob/1563ebbfa3f4c8530b334e194da9db630163041b/src/tests/fixtures/form.html#L103
[hotwired/turbo#272]: https://github.com/hotwired/turbo/issues/272#issuecomment-1198671657