Bugfix: Form Mode opt-in handle outside submitters
Prior to this commit, the `Session.formElementIsNavigatable` method
checked for the closest `<form>` element ancestor to its `element`
argument. Unfortunately, checking for an ancestor form is an incomplete
solution, since form controls can exist outside of the `<form>`
elsewhere in the document and reference the form through their
[`[form]`][form-attr] attributes.

This commit incorporates a check for that into the
`Session.formElementIsNavigatable` method. Since that method is only
ever called on a submission's `submitter` argument after it's been
verified to be present, this commit changes the signature to expect an
`element` (by omitting the `?` modifier) and renames the method to
`submitterIsNavigatable`.

[form-attr]: https://developer.mozilla.org/en-US/docs/Web/HTML/Element/Input#form