Dispatch `turbo:before-stream-render` with reference to `<turbo-stream>`
Since the `turbo:before-stream-render` event is dispatched _before_ the
`<turbo-stream>` element is connected to the document, it isn't possible
to dispatch the event with the element as its `Event.target` property.

To work around that constraint, dispatch the reference as
`CustomEvent.detail.newStream` (named in the same way as `newBody` for
`turbo:before-render` and `newFrame` for `turbo:before-frame-render`).