Add `[data-turbo-enforced]` to disable `data-turbo=false`
Developers can use [`data-turbo=false`](https://turbo.hotwired.dev/handbook/drive#disabling-turbo-drive-on-specific-links-or-forms) to disable Turbo Drive on specific links or form elements. But this introduces a problem in Turbo Native apps, where using `data-turbo=false` will exit the app and open the link in a web browser.

One of the best things about Turbo is that you can build features that will probably work natively without ever having to test natively. But `data-turbo=false` introduces an easy for developers to break this pact. A recent example I noticed in code review was a developer using `data-turbo=false` on a link to delete a record, because that would cause a full page reload and thus show an empty state again. (A better approach would be to use a Turbo Stream, or the CSS pattern described [here](https://boringrails.com/articles/css-tips-and-tricks-hotwire/).) On the native app, this link switched over the browser, so it didn't work at all.

In https://github.com/hotwired/turbo-site/pull/104 I added a warning about this, but warnings require people to read docs. So in this PR I'm proposing a way of controlling this using Turbo. This example shows what I propose:

```html
<div data-turbo="false">
  <a href="foo.html">Internal link. Full page navigation in browser; exits app in native.</a>
  <a href="https://www.example.com">External link. Full page navigation in browser; exits app in native.</a>
</div>

<div data-turbo-enforced>
  <div data-turbo="false">
    <a href="foo.html">Internal link. Link will be followed using Turbo! (data-turbo="false" is ignored)</a>
    <a href="https://www.example.com">External link. Full page navigation in browser; exits app in native.</a>
  </div>
</div>
```

The most common usage would be adding `data-turbo-enforced` to the `<body>` ([or relevant element](https://github.com/hotwired/turbo/pull/627)), either always or just when rendering in a native app (via user agent). I'd suggest doing it always so that you get the same behaviour in browser development as you do in app production.