Bugfix: Redirects and `[data-turbo-cache=false]`
When redirecting to a page that contains elements marked with
`[data-turbo-cache="false"]`, those elements are removed _before_ the
initial render, instead of _after_ the render and _before_ the page is
cached.

This behavior seems to have stemmed from [hotwired/turbo#516][], which
was shipped in response to [hotwired/turbo#515][].

As an alternative to the `willRender: false` option passed to
`this.adapter.visitProposedToLocation` in `Visit.followRedirect`, the
implementation can instead [rely on the presence of the
`turbo-frame[complete]`][comment] to guard against double fetching.

To guard against regressions, this commit adds coverage for the unwanted
behavior by redirecting from `navigation.html` to `cache_observer.html`,
and asserting the presence of a `[data-turbo-cache="false"]` element
that resembles and application's Flash messaging.

[hotwired/turbo#515]: https://github.com/hotwired/turbo/issues/515
[hotwired/turbo#516]: https://github.com/hotwired/turbo/pull/516
[comment]: https://github.com/hotwired/turbo/issues/515#issuecomment-1159738755
[hotwired/turbo#487]: https://github.com/hotwired/turbo/pull/487