Stream responses to GET should clear busy state
We set the `aria-busy` state on the document element whenever a link is
clicked that initiates a visit. After the visit completes, we clear this
state.

However, Turbo Stream responses are not visits, and so in the case where
the server chooses to respond with a stream response that busy state
wasn't being cleared.

To fix this, we'll now also clear the busy state when a stream response
is received in response to an HTTP request.