Test harness: Prevent infinite recusion
When serializing `CustomEvent.detail` objects from the browser to the
Playwright test harness, the current `serializeToChannel` implementation
is prone to recurse infinitely if an object nests another object that
refers to the outer object.

To prevent that, this commit tracks which objects have been visited
during the current serialization process, and avoid an infinitely
recursing loop.

Related to that change, this commit also add more descriptive types to
the `EventLog` and `MutationLog` arrays to clarify which positional
elements correspond to their serialized values.

Finally, extend the test server's `<turbo-stream>` creation actions to
support serializing an element with an `[id]` value, so that it can be
serialized from the browser to the test harness, and serve as a
`target.id` value.