`data-turbo-stream` links should not set aria-busy (#681)
Normally when following a link Turbo sets `[aria-busy="true"]` on the
document element. When the visit is complete, that attribute is removed.

However, when using `[data-turbo-stream]` to request a stream response
to a link, there is no visit to complete. That presented as a bug where
`aria-busy` was set at the start of the operation, but never cleared.

To solve this, we avoid setting the `aria-busy` attribute for links that
have a `data-turbo-stream` attribute.