Revert "Dispatch visit events on the initiator link or form (#695)"
The inclusion of the `initiator` in `VisitOptions` causes a JS crash in
the Turbo iOS adapter.

This is because Turbo iOS's implementation of `visitProposedToLocation`
relies on being able to pass the entire options struct to native code.
Initiator is an `Element`, which can't be passed in this way, resulting
in a `DataCloneError` exception.

There is some discussion on this [GitHub issue][]. We're reverting the
change for the moment, until we have an implementation of it that works
with all adapters.

[GitHub issue]: https://github.com/hotwired/turbo/issues/720

This reverts commit 887181796e8d5586de89c70327f31e9a81dcfec2.