CI: Restore passing tests
With the reverting of [hotwired/turbo#569][] (merged as
[hotwired/turbo#723][]), the "Visit with network error" test coverage
was missed in the diff.

Prior to the revert, the test was listening for
`turbo:fetch-request-error` events dispatched on the _link_ that was
clicked. Since that's no longer the case, this commit changes the test
to listen for `turbo:fetch-request-error` that bubble up to the
document.

[hotwired/turbo#569]: https://github.com/hotwired/turbo/pull/569
[hotwired/turbo#723]: https://github.com/hotwired/turbo/pull/723