Fix double `before-fetch-request` dispatch during reload
Closes https://github.com/hotwired/turbo/issues/739

Issue #739 mentions the issue at the root of the unexpected behavior:

> The reason is that `reload()` calls `this.removeAttribute("complete")`
> and the frame watches attribute changes and calls
> `this.loadSourceURL()` when the “complete” attribute is removed.

This commit resolves the issue in several steps.

First, add the `sourceURLReloaded()` callback to the
`FrameElementDelegate` interface.

Next, move the `FrameElement.reload()` implementation from
`FrameElement.reload()` to the `FrameController.sourceURLReloaded()`.

Finally, wrap the implementation's call to
`this.element.removeAttribute("complete")` within a
`ignoringChangesToAttribute("complete", () => ...)` block so that the
change to the attribute is temporarily ignored.