Cache PageSnapshot with original HTML from the frame's page
Before this change, when a promoted frame finished its navigation, a visit
was started and the visit would add a new PageSnapshot to the cache.
The problem is that this snapshot was looking at a page that was already
modified by the frame navigation, so the snapshot cache would be incorrect.

With this proposed change, we are taking a PageSnapshot as soon as the frame
navigation starts, so we guarantee that the page hasn't been modified yet.
Then, I'm passing this snapshot forward to the Visit and using it as the
correct cache instead of calculating a new one.