Resolve Issue #747
Closes https://github.com/hotwired/turbo/issues/747

Prior to this commit, `<turbo-frame>` elements navigated by
`[data-turbo-frame]` attributes were not promoted into Visits if
navigation initiator (a `<form>` element, for example) had another
`<turbo-frame>` as its ancestor.

To resolve that issue, this commit adds the
`proposeVisitIfNavigatedWithAction` method to the `FrameElementDelegate`
interface.

Instead of tracking _which_ `FrameElement` would navigate as a `Visit`
through an internal `frame?: FrameElement` property, the
`FrameController` methods **reach through** the `FrameElement` instance
that's resolved by the `FrameController.findFrameElement` method to
delegate navigation promotion to **that frame's delegate** instead of to
itself.