Check last rendered location for same-page visits
In [#595][] the logic for determining whether a visit is a "same page"
visit was changed. Previously we were comparing the new location with
the current location. In #595 we began comparing with the location of
the last visit that was started.

This works great when all the visits are web visits. However, it
introduced some problems with apps using the native adapters. In native
apps, navigations can be a combination of native screens and WebView
navigations, which means that we can't rely on the last visit having
been set.

This manifests as failing navigations in certain situations, such as
navigating back from a native modal to a WebView. It also caused some
problems restoring scroll positions when navigating back from one
WebView to another.

Restoring the original logic prevents these problems.

[#595]: https://github.com/hotwired/turbo/pull/595

This reverts commit 2e344b35d983fca636e064bc5ee1801c16908a04.