Remove custom `assert.equal` assertion
The bespoke `assert.equal` assertion override was introduced in order to
handle `.textContent` comparisons in an spacing-agnostic manner.

Unfortunately, its implementation was incorrect, which obscured several
failing tests throughout the file.

This commit removes that definition, and corrects several of the
test blocks to make assertions that account for the spacing differences
themselves.