Fix: Promoting lazy-loaded Frames
Prior to this commit, `<turbo-frame>` elements navigated via the
`AppearanceObserver` (powered by the [loading=lazy][] attribute) were
not accounting for their [data-turbo-action][] attributes.

To resolve that issue, this commit changes the `AppearanceObserver` and
`AppearanceObserverDelegate` to utilize a parameterized `<T extends
Element>` type so that the delegate callbacks can be invoked with
`FrameElement` references. With those changes in place, this commit
changes the `FrameController.elementAppearedInViewport` callback's
invocation to prepare a `PageSnapshot` and attempt to
`proposeVisitIfNavigatedWithAction`.

[loading=lazy]: https://turbo.hotwired.dev/reference/frames#lazy-loaded-frame
[data-turbo-action]: https://turbo.hotwired.dev/reference/frames#frame-that-promotes-navigations-to-visits