Re-use `getVisitAction` utility function
The logic for reading an `Action` from a list of `Element` instances'
`[data-turbo-action]` attribute is still scattered across several
modules, in spite of the `getVisitAction` declared in the `util` module.

This commit replaces duplicate implementations of the `getVisitAction`
method with the utility function.

With those calls added, the `isAction` type checking utility is only
ever invoked by `getVisitAction`, so this commit moves its
implementation to the `util` module, and removes all other invocations.