Skip Snapshot Caching for redirect visits
Closes [hotwired/turbo#794][]

Reverts the implementation change made in [hotwired/turbo#674][], and in
its place passes `shouldCacheSnapshot: false` alongside `willRender:
false` for the `action: "replace"` Visit proposed when following a
redirect.

In order to test this behavior, this commit introduces the
`readBodyMutationLogs` test utility function, along with the
`window.bodyMutationLogs` property and the `BodyMutationLog` type.

`BodyMutationLog` instances are pushed onto the log `Array` whenever the
`<body>` element is replaced.

[hotwired/turbo#794]: https://github.com/hotwired/turbo/issues/794
[hotwired/turbo#674]: https://github.com/hotwired/turbo/pull/674